+++ homeassistant/components/dyson/air_quality.py
@@ -6,7 +6,7 @@ from libpurecool.dyson_pure_state_v2 import DysonEnvironmentalSensorV2State
 
 from homeassistant.components.air_quality import DOMAIN, AirQualityEntity
 
-from . import DYSON_DEVICES
+from . import DYSON_DEVICES, DysonEntity
 
 ATTRIBUTION = "Dyson purifier air quality sensor"
 
@@ -29,7 +29,10 @@ def setup_platform(hass, config, add_entities, discovery_info=None):
     device_ids = [device.unique_id for device in hass.data[DYSON_AIQ_DEVICES]]
     new_entities = []
     for device in hass.data[DYSON_DEVICES]:
-        if isinstance(device, DysonPureCool) and device.serial not in device_ids:
+        if (
+            isinstance(device, DysonPureCool)
+            and f"air_quality/{device.serial}" not in device_ids
+        ):
             new_entities.append(DysonAirSensor(device))
 
     if not new_entities:
@@ -39,18 +42,13 @@ def setup_platform(hass, config, add_entities, discovery_info=None):
     add_entities(hass.data[DYSON_AIQ_DEVICES])
 
 
-class DysonAirSensor(AirQualityEntity):
+class DysonAirSensor(DysonEntity, AirQualityEntity):
     """Representation of a generic Dyson air quality sensor."""
 
     def __init__(self, device):
         """Create a new generic air quality Dyson sensor."""
-        self._device = device
+        super().__init__(device, "air_quality")
         self._old_value = None
-        self._name = device.name
-
-    async def async_added_to_hass(self):
-        """Call when entity is added to hass."""
-        self._device.add_message_listener(self.on_message)
 
     def on_message(self, message):
         """Handle new messages which are received from the fan."""
@@ -63,16 +61,6 @@ class DysonAirSensor(AirQualityEntity):
         ) and isinstance(message, DysonEnvironmentalSensorV2State):
             self._old_value = self._device.environmental_state
             self.schedule_update_ha_state()
-
-    @property
-    def should_poll(self):
-        """No polling needed."""
-        return False
-
-    @property
-    def name(self):
-        """Return the name of the Dyson sensor."""
-        return self._name
 
     @property
     def attribution(self):
@@ -116,11 +104,6 @@ class DysonAirSensor(AirQualityEntity):
         if self._device.environmental_state:
             return int(self._device.environmental_state.volatile_organic_compounds)
         return None
-
-    @property
-    def unique_id(self):
-        """Return the sensor's unique id."""
-        return self._device.serial
 
     @property
     def device_state_attributes(self):
